whatsapp_automation/requirements.txt
# FastAPI and Web
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
jinja2>=3.1.3

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# PostgreSQL
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Redis and Celery
redis==5.0.1
celery==5.3.6

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=42.0.0

# Rate limiting
slowapi>=0.1.9

# Logging
python-json-logger>=2.0.7

# OpenAI
openai==1.12.0

# HTTP Clients
httpx==0.26.0
requests>=2.31.0

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3
itsdangerous>=2.1.0
